cubes new constants
- restored i32 in WIDTH and HEIGHT in glb2 because in cubes1 they are similar, and removing the precision removed half the items

+++ examples/glb2/glb_renderpass/compute.js
@@ -29,7 +29,7 @@ fn main(in: ComputeIn) {
     let Y = y * (WORKGROUP_X * THREADS_X);
     let Z = z * (WORKGROUP_X * THREADS_X) * (WORKGROUP_Y * THREADS_Y);
 
-    let index = X + Y + Z;
+    let index = i32(X + Y + Z);
 
     let particle = &particles[index];
